+++ lizard_ocean/tests.py
@@ -11,3 +11,14 @@ class ConfTest(TestCase):
 
     def test_config_loaded(self):
         self.assertTrue(hasattr(settings, 'OCEAN_RASTER_BASEDIR'))
+
+
+class RasterTest(TestCase):
+
+    def test_get_rasters(self):
+        from .raster import get_rasters
+
+        rasters = get_rasters()
+        self.assertTrue('20130806' in rasters)
+        self.assertEquals(rasters['20130806'],
+                          settings.OCEAN_RASTER_BASEDIR + '/test_20130806.png')